// src/components/media-text-small-pattern/media-text-small-pattern.interfce.d.ts
import { HeaderLongProps } from '../header-long';
import { MediaTextSmallProps } from '../media-text-small/media-text-small.interface';
export interface MediaTextSmallPatternProps extends Partial<HeaderLongProps> {
    items: MediaTextSmallProps[];
    id?: string;
}
//# sourceMappingURL=media-text-small-pattern.interfce.d.ts.map